+++ src/lib.rs
@@ -17,6 +17,8 @@ Here are some of its properties:
   the language is quite possibly too slow where speed is the focus, but fast enough if
   you do any heavy lifting in Rust.
 - Main data structures are Vectors and Maps.
+- Builtin data structure pattern matchers and selectors which lead to a
+very powerful `match` operation.
 - No garbage collector. Memory and resource management relies only on reference counting and RAII.
   You can create your own drop functions.
 - Preserving Rust safety by not using `unsafe`.
@@ -289,6 +291,65 @@ std:assert_eq res1 :ok;
 std:assert_eq res1 :failed;
 ```
 
+## Builtin Structure Selectors
+
+Selectors work similar to XPath:
+`$S( *:{a=10} /b/1 )` first selects all maps from a vector,
+checks if they got a key-value pair that matches key=`a` and value=`10`.
+The selector path is walked for the matching maps and the `b` key
+is selected. Next the element at index `1` is selected and
+captured.
+
+```wlambda
+!struct = $[
+    ${ a = 10, b = $[ 1, 2, 3 ] },
+    ${ a = 10, b = $[ 4, 5, 6 ] },
+    ${ a = 20, b = $[ 8, 9,  20 ] },
+    ${ a = 20, b = $[ 8, 10, 30 ] },
+    ${ x = 99 },
+    ${ y = 99 },
+];
+
+? struct &> $S( *:{a=10} /b/1 ) {
+    std:assert_str_eq $\    $[2,5];
+} {
+    panic "Should've matched!";
+};
+```
+
+## Builting Structure Patterns
+
+A bit different but similar to the structure selectors `$S ...` are the `$P
+...` or `match` structure patterns:
+
+```wlambda
+!struct = $[
+    ${ a = 10, b = $[ 1, 2, 3 ] },
+    ${ a = 10, b = $[ 4, 5, 6 ] },
+    ${ a = 20, b = $[ 8, 9,  20 ] },
+    ${ a = 20, b = $[ 8, 10, 30 ] },
+    ${ x = 99 },
+    ${ y = 99 },
+];
+
+!res = $@vec iter elem struct {
+    $+ ~
+        match elem
+            ${ a = 10, b = childs }     => $[:childs_10, $\.childs]
+            ${ a = 20, b = childs }     => $[:childs_20, $\.childs]
+            :other;
+};
+
+std:assert_str_eq res $[
+    $[:childs_10,$[1, 2,   3]],
+    $[:childs_10,$[4, 5,   6]],
+    $[:childs_20,$[8, 9,  20]],
+    $[:childs_20,$[8, 10, 30]],
+    :other,
+    :other,
+];
+```
+
 ## Builtin (Regex) Pattern Matching
 
 ```wlambda
@@ -297,7 +358,7 @@ std:assert_eq res1 :failed;
 !crate  = $none;
 !domain = $none;
 
-? ($r{$^ (^$+[^:]) \:\/\/ (^$*[^/]) \/crates\/ (^$+[a-z]) } some_url) {
+? some_url <& $r{$^ (^$+[^:]) \:\/\/ (^$*[^/]) \/crates\/ (^$+[a-z]) } {
     .domain = $\.2;
     .crate = $\.3;
 };
@@ -445,13 +506,14 @@ assert_eq!(r.s(), "42");
 
 # Possible Roadmap
 
-There are several things that can be added more or less easily to
-WLambda. But I am currently working on making the language more
-complete for real world use. So my current goals are:
+Current remaining goals for WLambda are:
 
+- Fix remaining bugs.
+- Add missing standard library functions without dragging in more
+dependencies.
 - Improve and further document the VVal API for interacting with WLambda.
-- Improve reference documentation.
-- DONE: Add proper module support (via !@import and !@export).
+- Improve [WLambda Language Reference](https://docs.rs/wlambda/newest/wlambda/prelude/index.html#wlambda-reference) documentation.
+- DONE: Add proper module support (via `!@import` and `!@export`).
 - DONE: Add prototyped inheritance for OOP paradigm.
 - DONE: Add data structure matching/destructuring/selection primitives
 to the language.